another stupid mistake of mine, i blame python although its mostly my fault for not writing any tests

+++ lib/common.py
@@ -349,7 +349,7 @@ def weighted_order_choose(orders, n, feekey):
 	unless M < orderbook size, then phi goes up to the last order
 	'''
 	minfee = feekey(orders[0])
-	M = 1.5*n
+	M = int(1.5*n)
 	if len(orders) > M:
 		phi = feekey(orders[M]) - minfee
 	else:
